Actualización Tienda

+++ S4_TiendaSpring/src/main/java/es/rf/tienda/controller/CategoriaController.java
@@ -2,6 +2,7 @@ package es.rf.tienda.controller;
 
 
 import es.rf.tienda.objetos.repository.ICategoriaRepo;
+import es.rf.tienda.service.IServicioCategoria;
 
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -17,41 +18,45 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import es.rf.tienda.dominio.Categoria;
+import es.rf.tienda.exception.DAOException;
 
 @RestController
 @RequestMapping("/categorias")
 public class CategoriaController{
 	
 	@Autowired
-	private ICategoriaRepo cDao;
+	private IServicioCategoria cDao;
 	
 	@GetMapping("/{id}")
-	public String[] leerUno(@PathVariable("id") Integer id) {
-		try {
-			Categoria c = cDao.findById(id).get();
+	public String[] leerUno(@PathVariable("id") Integer id) throws DAOException {
+		if (cDao.idExist(id)) {
+			Categoria c = cDao.leerUno(id);
 			return new String[] {"200", c.toString()};
 		}
-		catch (NoSuchElementException e) {
+		else {
 			return new String[] {"400", "No existe registro solicitado"};	
 		}
 	}
 	
 	@GetMapping()
 	public List<Categoria> leerTodos(){
-		return cDao.findAll();
+		return cDao.leerTodos();
 	}
 
 	@PostMapping
 	public String[] insertar(@RequestBody Categoria c) {
 		c.setId_categoria(0);
-		cDao.save(c);
-		return new String[] {"200", "Registro guardado"};
+		if (cDao.insertar(c)) {
+			return new String[] {"200", "Registro insertado"};
+		}
+		else {
+			return new String[] {"400", "Registro no insertado"};
+		}
 	}
 	
 	@PutMapping
 	public String[] modificar(@RequestBody Categoria c) {
-		if (c.isValid()) {
-			cDao.save(c);
+		if (cDao.update(c)) {
 			return new String[] {"200", "Registro modificado"};
 		}
 		else {
@@ -60,13 +65,27 @@ public class CategoriaController{
 	}
 	
 	@DeleteMapping("/{id}")
-	public String[] eliminar(@PathVariable("id") Integer id) {
-		try {
+	public String[] eliminar(@PathVariable("id") Integer id) throws DAOException{
+		if (cDao.idExist(id)) {
 			cDao.deleteById(id);
 			return new String[] {"200", "Registro eliminado"};
+			
+			
 		}
-		catch (IllegalArgumentException e) {
-			return new String[] {"400", "No existe registro solicitado"};
+		else {
+			return new String[] {"400", "No existe registro solicitado para eliminar"};
+		}
+	}
+	
+	@DeleteMapping
+	public String[] eliminarCategorias(Categoria c) throws DAOException{
+		if (cDao.idExist(c.getId_categoria())) {
+			eliminar(c.getId_categoria());
+			return new String[] {"200", "Registro eliminado"};
+			
+		}
+		else {
+			return new String[] {"400", "No existe registro solicitado para eliminar"};
 		}
 	}
 }
